requirements.txt
BeautifulSoup4==4.6.0
requests==2.18.4
Flask==1.0.2
Pillow==5.1.0